refactor(eve): consolidate the Vercel project picker into pickProject
Inline pickExistingVercelProject into pickProject and drop the
vercel-project-picker module; move its tests into vercel-project.test.ts.
Adapted onto the modular rebase: VercelProjectListEntry is exported from
vercel-project-api, and the test capture helper accepts objects.
